make: move automake common bits to include file
This is a temporary solution to let us share automake settings globally.
An alternative this includes approach is to set flags globall in
configure.ac. Since many flags are platform and compiler dependent,
configure.ac is a good location for this, but also requires a lot more
work up front work. For now we want to keep the same settings while we
reorganize the build.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>