mon: minimal test for get_command_descriptions
The get_command_descriptions is not designed to be tested in C++ because
all the validation happens in ceph_argparse.py. Each test function calls
get_command_descriptions exactly once and does a trivial check on the
generated output. It does not really do more that code coverage.

With --debug-mon=20 each test function is expected to write the json
string to the standard output. The unit test can then be used as a
helper by other tests to generate various degenerated cases for test
purposes.

For instance:

  unittest_get_command_descriptions --gtest_filter=*.pull585 \
      --debug-mon=20

will generate a bogus json replicating the problem which was fixed
by https://github.com/ceph/ceph/pull/585

http://tracker.ceph.com/issues/6274 refs #6274

Signed-off-by: Loic Dachary <loic@dachary.org>